Final 1

+++ Shared/StatisticUpdate.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -13,19 +12,19 @@ namespace Shared
     public struct StatisticUpdate
     {
         public DateOnly Date { get; set; }
-        public int TotalSick { get; set; }
-        public int NewSick { get; set; }
-        public int TotalDead { get; set; }
-        public int NewDead { get; set; }
-        public int TotalVaccinated { get; set; }
-        public int NewVaccinated { get; set; }
+        public long TotalSick { get; set; }
+        public long NewSick { get; set; }
+        public long TotalDead { get; set; }
+        public long NewDead { get; set; }
+        public long TotalVaccinated { get; set; }
+        public long NewVaccinated { get; set; }
 
         public Dictionary<string, Region> RegionsByIso { get; set; } = new();
 
         /// <summary>
         /// Creates a statistic update record.
         /// </summary>
-        public StatisticUpdate(int newSick, int newDead, int newVaccinated, int totalSick, int totalDead, int totalVaccinated)
+        public StatisticUpdate(long newSick, long newDead, long newVaccinated, long totalSick, long totalDead, long totalVaccinated)
         {
             TotalSick = totalSick;
             TotalDead = totalDead;
@@ -34,7 +33,7 @@ namespace Shared
             NewDead = newDead;
             NewVaccinated = newVaccinated;
         }
-        public StatisticUpdate(DateOnly date, int newSick, int newDead, int newVaccinated, int totalSick, int totalDead, int totalVaccinated, Dictionary<string, Region> regionsByIso)
+        public StatisticUpdate(DateOnly date, long newSick, long newDead, long newVaccinated, long totalSick, long totalDead, long totalVaccinated, Dictionary<string, Region> regionsByIso)
         {
             Date = date;
             TotalSick = totalSick;
